feat: add publish prop, move theme definitions.
