+++ src/constants/index.ts
@@ -1,37 +1,3 @@
-import { ThemeDavids, ThemeDefault, ThemeDefaultPDF } from "@/theme";
-import { ThemeDefinition, ThemeName } from "@/types";
-
 export { SOCIAL_MEDIA_PLATFORMS } from "./social";
 
 export const titleSuffix = "| Amp'd Resume";
-
-export const themeDefinitions: Record<ThemeName, ThemeDefinition> = {
-  default: {
-    name: "Classic",
-    webComponent: ThemeDefault,
-    pdfComponent: ThemeDefaultPDF,
-    description: "The default theme for Amp'd Resume. Single-page resume with expanding sections.",
-    iconifyIcon: "fluent-emoji-flat:high-voltage",
-    authors: [
-      {
-        name: "Michael R. Dinerstein",
-        gitHubUrl: "https://github.com/missionmike",
-        linkedInUrl: "https://www.linkedin.com/in/michaeldinerstein/",
-      },
-    ],
-  },
-  davids: {
-    name: "David's Theme",
-    webComponent: ThemeDavids,
-    pdfComponent: null,
-    description: "The theme by David Schurer. Multi-page resume with a QR code to share.",
-    iconifyIcon: "fluent-emoji-flat:memo",
-    authors: [
-      {
-        name: "David Schurer",
-        gitHubUrl: "https://github.com/davidschurer",
-        linkedInUrl: "https://www.linkedin.com/in/david-schurer/",
-      },
-    ],
-  },
-};
